03/12/23 - Atualização de langs   - Verificação de pertencimento ao desativar / deletar carteira   - Ajustes ao salvar/deletar/desativar/criar carteira, padronizando e mantendo sempre em transações quando necessário

+++ app/Services/WalletService.php
@@ -6,11 +6,13 @@ use App\Http\Requests\CreateWalletRequest;
 use App\Models\User;
 use App\Models\UserWallet;
 use App\Models\Wallet;
+use Exception;
+use Illuminate\Support\Facades\DB;
+use Throwable;
 
 class WalletService
 {
     /**
-     * @param Wallet $wallet
      * @param User $user
      * @param CreateWalletRequest $request
      * @return Wallet
@@ -18,13 +20,15 @@ class WalletService
     public function save(User $user, CreateWalletRequest $request): Wallet
     {
         $wallet = new Wallet();
-        $wallet->name = $request->get('name');
-        $wallet->save();
+        DB::transaction(function () use ($user, $request, $wallet) {
+            $wallet->name = $request->get('name');
+            $wallet->save();
 
-        $userWallet = new UserWallet();
-        $userWallet->user_id = $user->id;
-        $userWallet->wallet_id = $wallet->id;
-        $userWallet->save();
+            $userWallet = new UserWallet();
+            $userWallet->user_id = $user->id;
+            $userWallet->wallet_id = $wallet->id;
+            $userWallet->save();
+        });
 
         return $wallet;
     }
@@ -33,20 +37,48 @@ class WalletService
      * @param User $user
      * @param Wallet $wallet
      * @return bool
+     * @throws Throwable
      */
     public function delete(User $user, Wallet $wallet): bool
     {
-        $userWallets = new UserWallet();
-        $userWallets = $userWallets->where([
-            'wallet_id' => $wallet->id
-        ])->with('wallet')->get();
-
-        foreach ($userWallets as $userWallet) {
-            $userWallet->where(['wallet_id' => $userWallet->wallet_id])->delete();
+        if (!$this->belongsToUser($wallet->id)) {
+            return false;
         }
 
-        $wallet->delete();
+        DB::transaction(function () use ($wallet) {
+            UserWallet::query()
+                ->where('wallet_id', '=', $wallet->id)
+                ->delete();
+
+            $wallet->delete();
+        });
 
         return true;
+    }
+
+    /**
+     * @throws Exception
+     * @throws Throwable
+     */
+    public function deactivate(Wallet $wallet): bool
+    {
+        if (!$this->belongsToUser($wallet->id)) {
+            return false;
+        }
+
+        $wallet->active = false;
+        return $wallet->saveOrFail();
+    }
+
+    /**
+     * @param int $walletId
+     * @return bool
+     */
+    public function belongsToUser(int $walletId): bool
+    {
+        return UserWallet::query()->where([
+            ['user_id', '=', auth()->id()],
+            ['wallet_id', '=', $walletId],
+        ])->exists();
     }
 }
